+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - ids parameter to transaction.query
 - ids parameter to transfer.query
+- hiddenFields parameter to boleto.pdf
+- ourNumber attribute to Boleto
 
 ## [0.6.0] - 2020-07-12
 ### Added